+++ src/hyp3_isce2/insar_tops_multi_burst.py
@@ -15,7 +15,7 @@ from hyp3_isce2 import packaging
 from hyp3_isce2.burst import validate_bursts
 from hyp3_isce2.insar_tops import insar_tops
 from hyp3_isce2.logger import configure_root_logger
-from hyp3_isce2.utils import get_multiburst_prefix, make_browse_image
+from hyp3_isce2.utils import get_publish_name, get_publish_prefix, make_browse_image
 
 
 gdal.UseExceptions()
@@ -161,6 +161,12 @@ def main():
         help='Additionally, publish products to this bucket. Necessary credentials must be provided '
         'via the `PUBLISH_ACCESS_KEY_ID` and `PUBLISH_SECRET_ACCESS_KEY` environment variables.',
     )
+    parser.add_argument(
+        '--publish-prefix',
+        type=str,
+        default=None,
+        help='Prefix for the bucket where the products will be published',
+    )
 
     args = parser.parse_args()
 
@@ -188,5 +194,9 @@ def main():
         packaging.upload_product_to_s3(product_dir, output_zip, args.bucket, args.bucket_prefix)
 
     if args.publish_bucket:
-        prefix, s3_name = get_multiburst_prefix(output_zip)
+        if args.publish_prefix:
+            prefix = args.publish_prefix
+        else:
+            prefix = get_publish_prefix(output_zip)
+        s3_name = get_publish_name(output_zip)
         packaging.upload_file_to_s3_with_publish_access_keys(output_zip, args.publish_bucket, prefix, s3_name)